feat : 예약 있는 테마 삭제 불가 예외 구현

+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -5,8 +5,12 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
+import roomescape.theme.exception.ThemeConstraintException;
 import roomescape.theme.exception.ThemeDuplicateException;
+import roomescape.theme.exception.ThemeErrorCode;
 import roomescape.theme.exception.ThemeNotFoundException;
 import roomescape.theme.repository.ThemeRepository;
 
@@ -16,6 +20,7 @@ import roomescape.theme.repository.ThemeRepository;
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
 
     @Transactional
     public Theme save(final String name, final String description, final String thumbnailUrl) {
@@ -30,6 +35,8 @@ public class ThemeService {
 
     @Transactional
     public void deleteById(final long themeId) {
+        if(reservationRepository.existsByThemeId(themeId))
+            throw new ThemeConstraintException();
         themeRepository.deleteById(themeId);
     }
